refactoring/ seperate noteCard from noteCard
